修改安全告警规范

+++ services/audio_service/client/src/renderer_in_client.cpp
@@ -1542,7 +1542,7 @@ int32_t RendererInClientInner::Write(uint8_t *pcmBuffer, size_t pcmBufferSize, u
     CHECK_AND_RETURN_RET_LOG(renderMode_ != RENDER_MODE_CALLBACK, ERR_INCORRECT_MODE,
         "Write with callback is not supported");
     int32_t ret = WriteInner(pcmBuffer, pcmBufferSize, metaBuffer, metaBufferSize);
-    return ret <= 0 ? ret : pcmBufferSize;
+    return ret <= 0 ? ret : static_cast<int32_t>(pcmBufferSize);
 }
 
 int32_t RendererInClientInner::Write(uint8_t *buffer, size_t bufferSize)
